requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
black==23.9.1
click==8.1.7
exceptiongroup==1.1.3
Faker==19.6.2
fastapi==0.103.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
protobuf==4.21.12
pydantic==1.10.13
pydantic_core==2.10.1
python-dateutil==2.8.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.41.1
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
uvicorn==0.23.2
